@@ -17,36 +17,44 @@
     "lint:markdown": "markdownlint .github/*.md && markdownlint README.md"
   },
   "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^6.1.1",
+    "@fortawesome/free-brands-svg-icons": "^6.1.1",
+    "@fortawesome/free-regular-svg-icons": "^6.1.1",
+    "@fortawesome/free-solid-svg-icons": "^6.1.1",
+    "@fortawesome/react-fontawesome": "^0.1.16",
     "compression": "^1.7.4",
-    "express": "^4.17.1",
+    "express": "^4.17.2",
     "morgan": "^1.10.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0",
+    "prop-types": "^15.7.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-router-dom": "^6.2.1",
     "serialize-javascript": "^6.0.0"
   },
   "devDependencies": {
-    "@babel/eslint-parser": "^7.16.0",
-    "@babel/preset-react": "^7.16.0",
-    "babel-preset-razzle": "^4.2.5",
-    "eslint": "^8.1.0",
+    "@babel/core": "^7.17.2",
+    "@babel/eslint-parser": "^7.17.0",
+    "@babel/preset-react": "^7.16.7",
+    "babel-preset-razzle": "^4.2.14",
+    "eslint": "^8.3.0",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.25.2",
+    "eslint-plugin-import": "^2.25.3",
     "eslint-plugin-jsx": "^0.1.0",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-promise": "^5.1.1",
-    "eslint-plugin-react": "^7.26.1",
+    "eslint-plugin-react": "^7.27.1",
     "html-webpack-plugin": "^5.5.0",
-    "markdownlint-cli": "^0.29.0",
-    "mini-css-extract-plugin": "^2.4.3",
-    "prettier": "^2.4.1",
-    "razzle": "^4.2.5",
-    "razzle-dev-utils": "^4.2.5",
-    "react-test-renderer": "^17.0.2",
-    "stylelint": "^14.0.1",
+    "markdownlint-cli": "^0.31.1",
+    "mini-css-extract-plugin": "^2.5.3",
+    "prettier": "^2.5.1",
+    "razzle": "^4.2.14",
+    "razzle-dev-utils": "^4.2.14",
+    "react-test-renderer": "^18.1.0",
+    "stylelint": "^14.9.0",
     "stylelint-config-prettier": "^9.0.3",
-    "webpack": "^5.61.0",
-    "webpack-dev-server": "^4.4.0"
+    "url": "^0.11.0",
+    "webpack": "^5.68.0",
+    "webpack-dev-server": "^4.9.2"
   }
 }
