@@ -1,12 +1,18 @@
 import React from 'react';
-import { Route, Switch } from 'react-router-dom';
+import { Route, Routes } from 'react-router-dom';
 import Home from '../Home/Home';
+import { library } from '@fortawesome/fontawesome-svg-core';
+import { fab } from '@fortawesome/free-brands-svg-icons';
+import { fas } from '@fortawesome/free-solid-svg-icons';
+import { far } from '@fortawesome/free-regular-svg-icons';
+
+library.add(fab, fas, far);
 
 const App = () => (
   <div className="container">
-    <Switch>
-      <Route exact path="/" component={Home} />
-    </Switch>
+    <Routes>
+      <Route exact path="/" element={<Home />} />
+    </Routes>
   </div>
 );
 
