@@ -1,6 +1,6 @@
 import App from './components/App/App';
 import React from 'react';
-import { StaticRouter } from 'react-router-dom';
+import { StaticRouter } from 'react-router-dom/server';
 import express from 'express';
 import { renderToString } from 'react-dom/server';
 import { runtimeConfig } from './config';
@@ -33,8 +33,24 @@ const jsScriptTagsFromAssets = (assets, entrypoint, extra = '') => {
 const theme = runtimeConfig.THEME === 'Dark' ? 'dark.css' : 'light.css';
 
 const server = express();
+
 if (process.env.NODE_ENV === 'production') {
-  server.use(morgan('combined'));
+  server.use(
+    morgan('combined', {
+      skip: (req, res) => {
+        if (
+          (req?.originalUrl?.includes('/healthcheck') ||
+            req?.baseUrl?.includes('/healthcheck')) &&
+          runtimeConfig?.SKIP_HEALTH_CHECK_LOGS &&
+          runtimeConfig.SKIP_HEALTH_CHECK_LOGS === 'true'
+        ) {
+          return true;
+        } else {
+          return false;
+        }
+      },
+    }),
+  );
   server.use(compression());
 }
 
@@ -61,10 +77,70 @@ server
         <title >${runtimeConfig.META_TITLE}</title>
         <meta name="description" content="${runtimeConfig.META_DESCRIPTION}">
         <meta name="author" content="${runtimeConfig.META_AUTHOR}">
+        <meta name="keywords" content="${runtimeConfig.META_KEYWORDS}">
         <meta name="robots" content="${
           runtimeConfig.META_INDEX_STATUS || 'noindex'
         }">
         <meta name="viewport" content="width=device-width, initial-scale=1">
+        ${
+          runtimeConfig.OG_SITE_NAME
+            ? `<meta property="og:site_name" content="${runtimeConfig.OG_SITE_NAME}" />`
+            : ''
+        }
+        ${
+          runtimeConfig.OG_TITLE
+            ? `<meta property="og:title" content="${runtimeConfig.OG_TITLE}" />`
+            : ''
+        }
+        ${
+          runtimeConfig.OG_DESCRIPTION
+            ? `<meta property="og:description" content="${runtimeConfig.OG_DESCRIPTION}" />`
+            : ''
+        }
+        ${
+          runtimeConfig.OG_URL
+            ? `<meta property="og:url" content="${runtimeConfig.OG_URL}" />`
+            : ''
+        }
+        <meta property="og:type" content="siteweb" />
+        ${
+          runtimeConfig.OG_IMAGE
+            ? `
+                <meta property="og:image" content="${runtimeConfig.OG_IMAGE}" />
+                <meta property="og:image:secure_url" content="${runtimeConfig.OG_IMAGE}" />
+            `
+            : ''
+        }
+        ${
+          runtimeConfig.OG_IMAGE_WIDTH
+            ? `<meta property="og:image:width" content="${runtimeConfig.OG_IMAGE_WIDTH}" />`
+            : ''
+        }
+        ${
+          runtimeConfig.OG_IMAGE_HEIGHT
+            ? `<meta property="og:image:height" content="${runtimeConfig.OG_IMAGE_HEIGHT}" />`
+            : ''
+        }
+        ${
+          runtimeConfig.TWITTER_CARD
+            ? `<meta property="twitter:card" content="${runtimeConfig.TWITTER_CARD}" />`
+            : ''
+        }
+        ${
+          runtimeConfig.TWITTER_IMAGE
+            ? `<meta property="twitter:image" content="${runtimeConfig.TWITTER_IMAGE}" />`
+            : ''
+        }
+        ${
+          runtimeConfig.TWITTER_SITE
+            ? `<meta property="twitter:site" content="${runtimeConfig.TWITTER_SITE}" />`
+            : ''
+        }
+        ${
+          runtimeConfig.TWITTER_CREATOR
+            ? `<meta property="twitter:creator" content="${runtimeConfig.TWITTER_CREATOR}" />`
+            : ''
+        }
         <link href="https://fonts.googleapis.com/css?family=Open+Sans:400,600,700,800&amp;display=swap" rel="stylesheet">
         <link rel="stylesheet" href="css/normalize.css">
         <link rel="stylesheet" href="css/${theme}">
@@ -93,9 +169,41 @@ server
             </script>`
             : ''
         }
+        ${
+          runtimeConfig.MATOMO_URL && runtimeConfig.MATOMO_SITE_ID
+            ? `
+            <!-- Matomo -->
+            <script type="text/javascript">
+                var _paq = window._paq || [];
+                /* tracker methods like "setCustomDimension" should be called before "trackPageView" */
+                _paq.push(['trackPageView']);
+                _paq.push(['enableLinkTracking']);
+                (function () {
+                    var u = "${runtimeConfig.MATOMO_URL}/";
+                    _paq.push(['setTrackerUrl', u + 'matomo.php']);
+                    _paq.push(['setSiteId', '${runtimeConfig.MATOMO_SITE_ID}']);
+                    var d = document, g = d.createElement('script'), s = d.getElementsByTagName('script')[0];
+                    g.type = 'text/javascript';
+                    g.async = true;
+                    g.defer = true;
+                    g.src = u + 'matomo.js';
+                    s.parentNode.insertBefore(g, s);
+                })();
+            </script>
+            <!-- Matomo End -->`
+            : ''
+        }
 
     </head>
     <body>
+        ${
+          runtimeConfig.MATOMO_URL && runtimeConfig.MATOMO_SITE_ID
+            ? `
+            <!-- Matomo Image Tracker-->
+            <img referrerpolicy="no-referrer-when-downgrade" src="${runtimeConfig.MATOMO_URL}/matomo.php?idsite=${runtimeConfig.MATOMO_SITE_ID}&amp;rec=1" style="border:0" alt="" />
+            <!-- End Matomo -->`
+            : ''
+        }
         <div id="root">${markup}</div>
         <script>window.env = ${serialize(runtimeConfig)};</script>
         ${jsScriptTagsFromAssets(assets, 'client', ' defer crossorigin')}
